CppServer/py/requirements.txt
colorama==0.4.4
Django==3.1.3
djangorestframework==3.12.2
